Module lin_inf::lin_graph

Tools for working with undirected graphs and their relation to formulae

Traits

LinGraph

Functions for working with graphs

Functions

edge_to_non_edge

Counts times an edge becomes a non-edge in an inference

is_medial_star

Given graphs p and q of size number_vars, determine whether p can be reduced to q using only the medial reduction.

is_module

Determines whether the sets of vertices xs and ys partitions the graph lg into two modules.

is_rewrite

Determines whether lg1 can be rewritten to lg2 (both of size n_vars) by rule lg1 -> lg2 (both of size n_vars_rule).

is_switch

Determine whether lg1 reduces to lg2 (both of which have size number_vars) by a single application of switch.

non_edge_to_edge

Counts times a non-edge becomes an edge in an inference

reduce_inference

Takes an inference of size number_vars and returns the least inference that can be obtained by applying a permutation to both the premise and conclusion. The returned inference is least with respect to a lexicographic ordering on first the premise and then the conclusion, based off the given ordering of graphs.