Function lin_inf::lin_graph::reduce_inference

pub fn reduce_inference<T: LinGraph + Ord>(
    premise: T, 
    conclusion: T, 
    number_vars: usize
) -> (T, T)

Takes an inference of size number_vars and returns the least inference that can be obtained by applying a permutation to both the premise and conclusion. The returned inference is least with respect to a lexicographic ordering on first the premise and then the conclusion, based off the given ordering of graphs.