Trait lin_inf::lin_graph::LinGraph

pub trait LinGraph: Sized + Eq + Send {

    fn get(&self, x: Node, y: Node) -> bool;
    fn build<T>(f: T, number_vars: usize) -> Self
    where
        T: Fn(Node, Node) -> bool;
    fn all_extensions(&self, number_vars: usize) -> Vec<Self> ;
    fn all_graphs(number_vars: usize) -> Vec<Self> ;

    fn p4_free(&self, number_vars: usize) -> bool { ... }
    fn all_p4_free(number_vars: usize) -> Vec<Self>  { ... }
    fn neighbours(&self, number_vars: usize, x: Node) -> HashSet<Node> { ... }
    fn apply_perm(&self, p: &Permutation) -> Self { ... }
    fn dualise(&self, number_vars: usize) -> Self { ... }
    fn max_cliques<M: MClique>(&self, number_vars: usize) -> Vec<M>  { ... }
    fn is_and_partition(&self, a: &Vec<Node>, b: &Vec<Node>) -> bool { ... }
    fn is_or_partition(&self, a: &Vec<Node>, b: &Vec<Node>) -> bool { ... }
    fn cograph_decomp(&self, number_vars: usize) -> Formula { ... }
    fn from_formula(f: &Formula) -> Self { ... }
    fn edges(&self, number_vars: usize) -> usize { ... }
    fn non_edges(&self, number_vars: usize) -> usize { ... }

}

Functions for working with graphs

Implementing this trait allows the type to act as an undirected graph, and equips it with functions for working with graphs and their relation to linear inferences. Note that graphs defined with this trait do not know their size and many of the functions require this size to be passed in. This is done to save memory space when working with large numbers of graphs.

Required Methods

fn get(&self, x: Node, y: Node) -> bool

Test whether there is an edge between nodes x and y Returns true if here is an edge between x and y and false otherwise. Note that this should return for all x and y, even if x and y are larger than the intended size of the graph

fn build<T>(f: T, number_vars: usize) -> Selfwhere
T: Fn(Node, Node) -> bool,

Build a graph of size number_vars from a function f describing the edge data.

For all x and y less than number_vars we should have:

!assert_eq(build(f,number_vars).get(x,y),f(x,y))

and if x or y are greater than or equal to number_vars then we should have:

!assert_eq(build(f,number-vars).get(x,y), false)

fn all_extensions(&self, number_vars: usize) -> Vec<Self>

Returns all possible extensions of a number_vars sized graph to a number_vars + 1 sized graph.

fn all_graphs(number_vars: usize) -> Vec<Self>

Returns all graphs of size number_vars

Provided Methods

fn p4_free(&self, number_vars: usize) -> bool

Decides if a graph is P4 free

Returns true if none of the induced subgraphs of the input graph are isomorphic to P4

fn all_p4_free(number_vars: usize) -> Vec<Self>

Returns all P4 free graphs of size number_vars.

fn neighbours(&self, number_vars: usize, x: Node) -> HashSet<Node>

Finds all nodes adjacent to node x in the given graph of size number_vars

fn apply_perm(&self, p: &Permutation) -> Self

Apply a permutation to a graph to obtain a new graph.

fn dualise(&self, number_vars: usize) -> Self

Fully dualise a graph of size number_vars (invert every edge to a non edge and every non edge to an edge)

fn max_cliques<M: MClique>(&self, number_vars: usize) -> Vec<M>

Return all max cliques of a graph

fn is_and_partition(&self, a: &Vec<Node>, b: &Vec<Node>) -> bool

Determines whether a graph is the join of the induced subgraphs with vertices a and b.

fn is_or_partition(&self, a: &Vec<Node>, b: &Vec<Node>) -> bool

Determines whether a graph is the disjoint union of the induced subgraphs with vertices a and b.

fn cograph_decomp(&self, number_vars: usize) -> Formula

Finds the cograph decomposition of a P4-free graph of size number_vars, returning a formula. This function does not check that the graph is P4-free, and will fail if it is not.

fn from_formula(f: &Formula) -> Self

Builds a graph from a formula f.

fn edges(&self, number_vars: usize) -> usize

Counts number of edges.

fn non_edges(&self, number_vars: usize) -> usize

Counts number of non-edges.