Trait lin_inf::mclique::MClique

pub trait MClique {
    fn mc_from_set(set: &HashSet<Node>) -> Self;
    fn mc_contains(&self, x: Node) -> bool;
    fn mc_subset(&self, x: &Self) -> bool;
    fn mc_remove(&self, x: Node) -> Self;
}

Trait acting as a wrapper around set like structures exposing the functions needed for representing a maximum clique of a graph.

Required Methods

fn mc_from_set(set: &HashSet<Node>) -> Self

A maximum clique is a set of nodes and can be build from a HashSet.

fn mc_contains(&self, x: Node) -> bool

Check if a node is contained in a maximum clique.

fn mc_subset(&self, x: &Self) -> bool

Check if a maximum clique is a subset of another maximum clique x.

fn mc_remove(&self, x: Node) -> Self

Remove a node x from a maximum clique.